🚚 Move notification components out of UI folder
