+++ frontend/src/components/calendar/notifications/NotificationCard/NotificationCard.scss
@@ -1,4 +1,4 @@
-@import '../../../App.scss';
+@import '../../../../App.scss';
 
 .card {
   padding: var(--sp-base) var(--sp-lg);
